fixed data structure being a project and moved tests to the proper path
