--- python/data_strtucture/trees/pyproject.toml
@@ -1,17 +1,0 @@
-[tool.poetry]
-name = "trees"
-version = "0.1.0"
-description = ""
-authors = ["MusaabShalaldeh <musaabsh65@yahoo.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.9"
-black = "^21.9b0"
-flake8 = "^4.0.1"
-
-[tool.poetry.dev-dependencies]
-pytest = "^5.2"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
